+++ src/domain/entities/resonatorStats.js
@@ -1,5 +1,5 @@
 import { v4 as uuid } from "uuid";
-import log from '../../infra/log';
+import log from '../../logging/log';
 
 export default class ResonatorStats {
     constructor({
