+++ src/server.js
@@ -1,6 +1,6 @@
 import api from "./api";
 import cfg from "./cfg";
-import initInfra from "./infra";
+import initInfra from "./logging";
 import emailSchedulingJob from "./emailScheduler";
 import { setVapidKeys } from "./emailScheduler/push";
 // import syncCalendarsJob from "./calendars/calendarsSync";
